+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### added
 * record 可以直接处理内存中的 DataFrame 账单。
+* ioconf 增加自定义 key_func, 兼容缓存层对 key 支持不够的情况 （大小写不敏感，特殊字符等）
 
 ### fixed
 * evaluate 类的可视化函数补充 rendered 选项